requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.7
cssselect2==0.8.0
ecdsa==0.19.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httptools==0.6.4
idna==3.10
kombu==5.5.4
lxml==6.0.1
oscrypto==1.3.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pypdf==6.0.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
svglib==1.5.1
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
